requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
cachetools==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.3
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
distro==1.9.0
Django==4.2.9
et-xmlfile==1.1.0
exceptiongroup==1.2.0
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
image==1.5.33
importlib-metadata==7.0.1
Jinja2==3.1.3
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaitaistruct==0.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
measures==1.4.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
narwhals==1.4.0
networkx==3.2.1
numpy==1.26.3
openai==1.33.0
opencv-python==4.9.0.80
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
pefile==2023.2.7
pillow==10.4.0
Pillow-PIL==0.1.dev0
protobuf==5.27.3
pyarrow==17.0.0
pyasn1==0.6.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
pydivert==2.1.0
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyMsgBox==1.0.9
pymssql==2.2.4
pyodbc==5.0.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.0.7
pytz==2024.1
pywin32-ctypes==0.2.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
selenium==4.23.1
selenium-wire==5.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stem==1.8.2
streamlit==1.37.1
sympy==1.12
tenacity==8.5.0
toml==0.10.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.1
watchdog==4.0.2
webdriver-manager==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
